+++ src/main/kotlin/com/zepben/evolve/database/sqlite/tables/iec61970/base/auxiliaryequipment/TablePotentialTransformers.kt
@@ -9,13 +9,14 @@
 package com.zepben.evolve.database.sqlite.tables.iec61970.base.auxiliaryequipment
 
 import com.zepben.evolve.database.sqlite.tables.Column
+import com.zepben.evolve.database.sqlite.tables.Column.Nullable.NOT_NULL
 import com.zepben.evolve.database.sqlite.tables.Column.Nullable.NULL
 
 @Suppress("PropertyName")
 class TablePotentialTransformers : TableSensors() {
 
     val POTENTIAL_TRANSFORMER_INFO_MRID = Column(++columnIndex, "potential_transformer_info_mrid", "TEXT", NULL)
-    val TYPE = Column(++columnIndex, "type", "TEXT", NULL)
+    val TYPE = Column(++columnIndex, "type", "TEXT", NOT_NULL)
 
     override fun name(): String {
         return "potential_transformers"